requirements.txt
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
docker==4.1.0
idna==2.8
importlib-metadata==0.23
jsonschema==3.1.1
more-itertools==7.2.0
psutil==5.6.3
pyrsistent==0.15.4
requests==2.22.0
six==1.12.0
urllib3==1.25.6
websocket-client==0.56.0
zipp==0.6.0
